+++ functions.php
@@ -1,15 +1,5 @@
   <?php
 
-  // Build out URI to reload from form dropdown
-  // Need full url for this to work in Opera Mini
-  $pageURL = (@$_SERVER["HTTPS"] == "on") ? "https://" : "http://";
-
-  if (isset($_POST['sg_uri']) && isset($_POST['sg_section_switcher'])) {
-     $pageURL .= $_POST[sg_uri].$_POST[sg_section_switcher];
-     $pageURL = htmlspecialchars( filter_var( $pageURL, FILTER_SANITIZE_URL ) );
-     header("Location: $pageURL");
-  }
-
   // Display title of each markup samples as a select option
   function listMarkupAsOptions ($type) {
     $files = array();