+++ index.php
@@ -15,25 +15,6 @@
 <a href="#main" class="sg-visually-hidden sg-visually-hidden-focusable">Skip to main content</a>
 <div id="top" class="sg-header sg-container">
   <h1 class="sg-logo">STYLE GUIDE <span>BOILERPLATE</span></h1>
-  <form id="js-sg-nav" action=""  method="post" class="sg-nav">
-    <label for="js-sg-section-switcher" class="sg-visually-hidden">Navigation</label>
-    <select id="js-sg-section-switcher" class="sg-section-switcher" name="sg_section_switcher">
-        <option value="">Jump To Section:</option>
-        <optgroup label="Intro">
-          <option value="#sg-about">About</option>
-          <option value="#sg-colors">Colors</option>
-          <option value="#sg-fontStacks">Font-Stacks</option>
-        </optgroup>
-        <optgroup label="Base HTML">
-          <?php listMarkupAsOptions('base'); ?>
-        </optgroup>
-        <optgroup label="Patterns">
-          <?php listMarkupAsOptions('patterns'); ?>
-        </optgroup>
-    </select>
-    <input type="hidden" name="sg_uri" value="<?php echo $_SERVER["SERVER_NAME"].$_SERVER["REQUEST_URI"]; ?>">
-    <button type="submit" class="sg-submit-btn">Go</button>
-  </form><!--/.sg-nav-->
 </div><!--/.sg-header-->
 
 <div id="main" class="sg-body sg-container">